Fixes broken tests in returns matching algorithm.
