+++ test/modules/billing/services/two-part-tariff-service/sample-data/sample-matching-data-1.js
@@ -1,130 +1,110 @@
-const chargeVersion = {
-  chargeVersionId: 'charge-version-1',
-  licenceRef: 'licence-ref',
-  scheme: 'alcs',
-  versionNumber: 104,
-  startDate: '2016-07-28T23:00:00.000Z',
-  status: 'superseded',
-  apportionment: false,
-  error: false,
-  endDate: '2018-04-07T23:00:00.000Z',
-  billedUptoDate: '2018-04-07T23:00:00.000Z',
-  dateCreated: '2019-09-26T08:12:02.492Z',
-  dateUpdated: '2019-09-26T08:12:02.492Z',
-  source: 'nald',
-  chargeElements: [{
-    chargeElementId: 'charge-element-1',
-    chargeVersionId: 'charge-version-1',
-    abstractionPeriodStartDay: 1,
-    abstractionPeriodStartMonth: 3,
-    abstractionPeriodEndDay: 30,
-    abstractionPeriodEndMonth: 9,
-    authorisedAnnualQuantity: '91.4',
-    season: 'summer',
-    seasonDerived: 'summer',
-    source: 'unsupported',
-    loss: 'high',
-    purposePrimary: 'A',
-    purposeSecondary: 'AGR',
-    purposeTertiary: '420',
-    factorsOverridden: false,
-    billableAnnualQuantity: '91.4',
-    timeLimitedStartDate: '2008-03-31T23:00:00.000Z',
-    timeLimitedEndDate: '2018-03-30T23:00:00.000Z',
-    description: 'T/L B/H ABSTRACTION',
-    srocCategory: null,
-    dateCreated: '2019-09-26T08:12:08.346Z',
-    dateUpdated: '2019-09-26T08:12:08.346Z',
-    purposePrimaryDescription: 'Agriculture',
-    purposeSecondaryDescription: 'General Agriculture',
-    purposeTertiaryDescription: 'Spray Irrigation - Storage',
-    eiucSource: 'other',
-    chargeAgreements: [{
-      chargeAgreementId: 'charge-agreement-1',
-      chargeElementId: 'charge-element-1',
-      agreementCode: 'S127',
-      startDate: '2008-03-31T23:00:00.000Z',
-      endDate: null,
-      signedDate: null,
-      fileReference: null,
-      description: null,
-      dateCreated: '2019-09-26T08:12:17.107Z',
-      dateUpdated: '2019-09-26T08:12:17.107Z',
-      agreementDescription: 'Section 127 (Two Part Tariff)'
-    }]
+const chargeElements = [{
+  id: 'charge-element-1',
+  abstractionPeriod: {
+    startDay: 1,
+    startMonth: 3,
+    endDay: 30,
+    endMonth: 9
   },
-  {
-    chargeElementId: 'charge-element-2',
-    chargeVersionId: 'charge-version-1',
-    abstractionPeriodStartDay: 1,
-    abstractionPeriodStartMonth: 3,
-    abstractionPeriodEndDay: 30,
-    abstractionPeriodEndMonth: 9,
-    authorisedAnnualQuantity: '22.7',
-    season: 'summer',
-    seasonDerived: 'summer',
-    source: 'unsupported',
-    loss: 'high',
-    purposePrimary: 'A',
-    purposeSecondary: 'AGR',
-    purposeTertiary: '420',
-    factorsOverridden: false,
-    billableAnnualQuantity: '22.7',
-    timeLimitedStartDate: null,
-    timeLimitedEndDate: null,
-    description: '2 BORES-NON TL ABSTR',
-    srocCategory: null,
-    dateCreated: '2019-09-26T08:12:08.346Z',
-    dateUpdated: '2019-09-26T08:12:08.346Z',
-    purposePrimaryDescription: 'Agriculture',
-    purposeSecondaryDescription: 'General Agriculture',
-    purposeTertiaryDescription: 'Spray Irrigation - Storage',
-    eiucSource: 'other',
-    chargeAgreements: [{
-      chargeAgreementId: 'charge-agreement-2',
-      chargeElementId: 'charge-element-2',
-      agreementCode: 'S127',
-      startDate: '2004-11-19T00:00:00.000Z',
-      endDate: null,
-      signedDate: null,
-      fileReference: null,
-      description: null,
-      dateCreated: '2019-09-26T08:12:17.107Z',
-      dateUpdated: '2019-09-26T08:12:17.107Z',
-      agreementDescription: 'Section 127 (Two Part Tariff)'
-    }]
+  authorisedAnnualQuantity: '91.4',
+  season: 'summer',
+  source: 'unsupported',
+  loss: 'high',
+  purposeUse: {
+    type: 'use',
+    code: '420',
+    name: 'Spray Irrigation - Storage'
   },
-  {
-    chargeElementId: 'charge-element-3',
-    chargeVersionId: 'charge-version-1',
-    abstractionPeriodStartDay: 1,
-    abstractionPeriodStartMonth: 1,
-    abstractionPeriodEndDay: 31,
-    abstractionPeriodEndMonth: 12,
-    authorisedAnnualQuantity: '4.5',
-    season: 'all year',
-    seasonDerived: 'all year',
-    source: 'unsupported',
-    loss: 'medium',
-    purposePrimary: 'A',
-    purposeSecondary: 'AGR',
-    purposeTertiary: '140',
-    factorsOverridden: false,
-    billableAnnualQuantity: '4.5',
-    timeLimitedStartDate: null,
-    timeLimitedEndDate: null,
-    description: 'BOREHOLES FOR GENERAL AGRICULTURE',
-    srocCategory: null,
-    dateCreated: '2019-09-26T08:12:08.346Z',
-    dateUpdated: '2019-09-26T08:12:08.346Z',
-    purposePrimaryDescription: 'Agriculture',
-    purposeSecondaryDescription: 'General Agriculture',
-    purposeTertiaryDescription: 'General Farming & Domestic',
-    eiucSource: 'other',
-    chargeAgreements: []
-  }
-  ]
-};
+  factorsOverridden: false,
+  billableAnnualQuantity: '91.4',
+  timeLimitedPeriod: {
+    startDate: '2008-03-31T23:00:00.000Z',
+    endDate: '2018-03-30T23:00:00.000Z'
+  },
+  description: 'T/L B/H ABSTRACTION',
+  srocCategory: null,
+  dateCreated: '2019-09-26T08:12:08.346Z',
+  dateUpdated: '2019-09-26T08:12:08.346Z',
+  eiucSource: 'other',
+  chargeAgreements: [{
+    chargeAgreementId: 'charge-agreement-1',
+    id: 'charge-element-1',
+    agreementCode: 'S127',
+    startDate: '2008-03-31T23:00:00.000Z',
+    endDate: null,
+    signedDate: null,
+    fileReference: null,
+    description: null,
+    dateCreated: '2019-09-26T08:12:17.107Z',
+    dateUpdated: '2019-09-26T08:12:17.107Z',
+    agreementDescription: 'Section 127 (Two Part Tariff)'
+  }]
+},
+{
+  id: 'charge-element-2',
+  abstractionPeriod: {
+    startDay: 1,
+    startMonth: 3,
+    endDay: 30,
+    endMonth: 9
+  },
+  authorisedAnnualQuantity: '22.7',
+  season: 'summer',
+  source: 'unsupported',
+  loss: 'high',
+  purposeUse: {
+    type: 'use',
+    code: '420',
+    name: 'Spray Irrigation - Storage'
+  },
+  factorsOverridden: false,
+  billableAnnualQuantity: '22.7',
+  description: '2 BORES-NON TL ABSTR',
+  srocCategory: null,
+  dateCreated: '2019-09-26T08:12:08.346Z',
+  dateUpdated: '2019-09-26T08:12:08.346Z',
+  eiucSource: 'other',
+  chargeAgreements: [{
+    chargeAgreementId: 'charge-agreement-2',
+    id: 'charge-element-2',
+    agreementCode: 'S127',
+    startDate: '2004-11-19T00:00:00.000Z',
+    endDate: null,
+    signedDate: null,
+    fileReference: null,
+    description: null,
+    dateCreated: '2019-09-26T08:12:17.107Z',
+    dateUpdated: '2019-09-26T08:12:17.107Z',
+    agreementDescription: 'Section 127 (Two Part Tariff)'
+  }]
+},
+{
+  id: 'charge-element-3',
+  abstractionPeriod: {
+    startDay: 1,
+    startMonth: 1,
+    endDay: 31,
+    endMonth: 12
+  },
+  authorisedAnnualQuantity: '4.5',
+  season: 'all year',
+  seasonDerived: 'all year',
+  source: 'unsupported',
+  loss: 'medium',
+  purposeUse: {
+    type: 'use',
+    code: '140',
+    name: 'General Farming & Domestic'
+  },
+  factorsOverridden: false,
+  billableAnnualQuantity: '4.5',
+  description: 'BOREHOLES FOR GENERAL AGRICULTURE',
+  srocCategory: null,
+  dateCreated: '2019-09-26T08:12:08.346Z',
+  dateUpdated: '2019-09-26T08:12:08.346Z',
+  eiucSource: 'other',
+  chargeAgreements: []
+}];
 
 const returns = [{
   returnId: 'return-1',
@@ -432,6 +412,6 @@ const returns = [{
 }];
 
 module.exports = {
-  chargeVersion,
+  chargeElements,
   returns
 };
