+++ test/modules/billing/services/two-part-tariff-service/sample-data/sample-matching-data-3.js
@@ -1,172 +1,148 @@
-const chargeVersion = {
-  chargeVersionId: 'charge-version-1',
-  licenceRef: 'licence-ref',
-  scheme: 'alcs',
-  versionNumber: 105,
-  startDate: '2017-10-11T23:00:00.000Z',
-  status: 'current',
-  apportionment: false,
-  error: false,
-  endDate: null,
-  billedUptoDate: '2020-03-30T23:00:00.000Z',
-  dateCreated: '2019-09-26T08:12:02.492Z',
-  dateUpdated: '2019-09-26T08:12:02.492Z',
-  source: 'nald',
-  chargeElements: [
-    {
-      chargeElementId: 'charge-element-1',
-      chargeVersionId: 'charge-version-1',
-      abstractionPeriodStartDay: 1,
-      abstractionPeriodStartMonth: 11,
-      abstractionPeriodEndDay: 31,
-      abstractionPeriodEndMonth: 3,
-      authorisedAnnualQuantity: '82',
-      season: 'winter',
-      seasonDerived: 'winter',
-      source: 'unsupported',
-      loss: 'high',
-      purposePrimary: 'A',
-      purposeSecondary: 'AGR',
-      purposeTertiary: '420',
-      factorsOverridden: false,
-      billableAnnualQuantity: null,
-      timeLimitedStartDate: '2017-10-11T23:00:00.000Z',
-      timeLimitedEndDate: '2030-03-31T00:00:00.000Z',
-      description: 'River Oak',
-      srocCategory: null,
-      dateCreated: '2019-09-26T08:12:08.346Z',
-      dateUpdated: '2019-09-26T08:12:08.346Z',
-      purposePrimaryDescription: 'Agriculture',
-      purposeSecondaryDescription: 'General Agriculture',
-      purposeTertiaryDescription: 'Spray Irrigation - Storage',
-      eiucSource: 'other',
-      chargeAgreements: [{
-        chargeAgreementId: 'charge-agreement-1',
-        chargeElementId: 'charge-element-1',
-        agreementCode: 'S127',
-        startDate: '2017-10-11T23:00:00.000Z',
-        endDate: null,
-        signedDate: null,
-        fileReference: null,
-        description: null,
-        dateCreated: '2019-09-26T08:12:17.107Z',
-        dateUpdated: '2019-09-26T08:12:17.107Z',
-        agreementDescription: 'Section 127 (Two Part Tariff)'
-      }]
+const chargeElements = [
+  {
+    id: 'charge-element-1',
+    abstractionPeriod: {
+      startDay: 1,
+      startMonth: 11,
+      endDay: 31,
+      endMonth: 3
     },
-    {
-      chargeElementId: 'charge-element-2',
-      chargeVersionId: 'charge-version-1',
-      abstractionPeriodStartDay: 1,
-      abstractionPeriodStartMonth: 4,
-      abstractionPeriodEndDay: 31,
-      abstractionPeriodEndMonth: 10,
-      authorisedAnnualQuantity: '22.7',
-      season: 'summer',
-      seasonDerived: 'summer',
-      source: 'unsupported',
-      loss: 'high',
-      purposePrimary: 'A',
-      purposeSecondary: 'AGR',
-      purposeTertiary: '400',
-      factorsOverridden: false,
-      billableAnnualQuantity: '2.7',
-      timeLimitedStartDate: null,
-      timeLimitedEndDate: null,
-      description: 'Reservoir',
-      srocCategory: null,
-      dateCreated: '2019-09-26T08:12:08.346Z',
-      dateUpdated: '2019-09-26T08:12:08.346Z',
-      purposePrimaryDescription: 'Agriculture',
-      purposeSecondaryDescription: 'General Agriculture',
-      purposeTertiaryDescription: 'Spray Irrigation - Direct',
-      eiucSource: 'other',
-      chargeAgreements: [{
-        chargeAgreementId: 'charge-agreement-2',
-        chargeElementId: 'charge-element-2',
-        agreementCode: 'S127',
-        startDate: '2017-10-11T23:00:00.000Z',
-        endDate: null,
-        signedDate: null,
-        fileReference: null,
-        description: null,
-        dateCreated: '2019-09-26T08:12:17.107Z',
-        dateUpdated: '2019-09-26T08:12:17.107Z',
-        agreementDescription: 'Section 127 (Two Part Tariff)'
-      }]
+    authorisedAnnualQuantity: '82',
+    season: 'winter',
+    source: 'unsupported',
+    loss: 'high',
+    purposeUse: {
+      type: 'use',
+      code: '420',
+      name: 'Spray Irrigation - Storage'
     },
-    {
-      chargeElementId: 'charge-element-3',
-      chargeVersionId: 'charge-version-1',
-      abstractionPeriodStartDay: 1,
-      abstractionPeriodStartMonth: 11,
-      abstractionPeriodEndDay: 31,
-      abstractionPeriodEndMonth: 3,
-      authorisedAnnualQuantity: '20',
-      season: 'winter',
-      seasonDerived: 'winter',
-      source: 'unsupported',
-      loss: 'very low',
-      purposePrimary: 'A',
-      purposeSecondary: 'AGR',
-      purposeTertiary: '650',
-      factorsOverridden: false,
-      billableAnnualQuantity: null,
-      timeLimitedStartDate: null,
-      timeLimitedEndDate: null,
-      description: 'River Oak',
-      srocCategory: null,
-      dateCreated: '2019-09-26T08:12:08.346Z',
-      dateUpdated: '2019-09-26T08:12:08.346Z',
-      purposePrimaryDescription: 'Agriculture',
-      purposeSecondaryDescription: 'General Agriculture',
-      purposeTertiaryDescription: 'Transfer Between Sources (Post Water Act 2003)',
-      eiucSource: 'other',
-      chargeAgreements: []
+    factorsOverridden: false,
+    billableAnnualQuantity: null,
+    timeLimitedPeriod: {
+      startDate: '2017-10-11T23:00:00.000Z',
+      endDate: '2030-03-31T00:00:00.000Z'
     },
-    {
-      chargeElementId: 'charge-element-4',
-      chargeVersionId: 'charge-version-1',
-      abstractionPeriodStartDay: 1,
-      abstractionPeriodStartMonth: 4,
-      abstractionPeriodEndDay: 31,
-      abstractionPeriodEndMonth: 10,
-      authorisedAnnualQuantity: '22.7',
-      season: 'winter',
-      seasonDerived: 'summer',
-      source: 'unsupported',
-      loss: 'high',
-      purposePrimary: 'A',
-      purposeSecondary: 'AGR',
-      purposeTertiary: '400',
-      factorsOverridden: true,
-      billableAnnualQuantity: '20',
-      timeLimitedStartDate: null,
-      timeLimitedEndDate: null,
-      description: 'Reservoir',
-      srocCategory: null,
-      dateCreated: '2019-09-26T08:12:08.346Z',
-      dateUpdated: '2019-09-26T08:12:08.346Z',
-      purposePrimaryDescription: 'Agriculture',
-      purposeSecondaryDescription: 'General Agriculture',
-      purposeTertiaryDescription: 'Spray Irrigation - Direct',
-      eiucSource: 'other',
-      chargeAgreements: [{
-        chargeAgreementId: 'charge-agreement-3',
-        chargeElementId: 'charge-element-4',
-        agreementCode: 'S127',
-        startDate: '2017-10-11T23:00:00.000Z',
-        endDate: null,
-        signedDate: null,
-        fileReference: null,
-        description: null,
-        dateCreated: '2019-09-26T08:12:17.107Z',
-        dateUpdated: '2019-09-26T08:12:17.107Z',
-        agreementDescription: 'Section 127 (Two Part Tariff)'
-      }]
-    }
-  ]
-};
+    description: 'River Oak',
+    srocCategory: null,
+    dateCreated: '2019-09-26T08:12:08.346Z',
+    dateUpdated: '2019-09-26T08:12:08.346Z',
+    eiucSource: 'other',
+    chargeAgreements: [{
+      chargeAgreementId: 'charge-agreement-1',
+      id: 'charge-element-1',
+      agreementCode: 'S127',
+      startDate: '2017-10-11T23:00:00.000Z',
+      endDate: null,
+      signedDate: null,
+      fileReference: null,
+      description: null,
+      dateCreated: '2019-09-26T08:12:17.107Z',
+      dateUpdated: '2019-09-26T08:12:17.107Z',
+      agreementDescription: 'Section 127 (Two Part Tariff)'
+    }]
+  },
+  {
+    id: 'charge-element-2',
+    abstractionPeriod: {
+      startDay: 1,
+      startMonth: 4,
+      endDay: 31,
+      endMonth: 10
+    },
+    authorisedAnnualQuantity: '22.7',
+    season: 'summer',
+    source: 'unsupported',
+    loss: 'high',
+    purposeUse: {
+      type: 'use',
+      code: '400',
+      name: 'Spray Irrigation - Direct'
+    },
+    factorsOverridden: false,
+    billableAnnualQuantity: '2.7',
+    description: 'Reservoir',
+    srocCategory: null,
+    dateCreated: '2019-09-26T08:12:08.346Z',
+    dateUpdated: '2019-09-26T08:12:08.346Z',
+    eiucSource: 'other',
+    chargeAgreements: [{
+      chargeAgreementId: 'charge-agreement-2',
+      id: 'charge-element-2',
+      agreementCode: 'S127',
+      startDate: '2017-10-11T23:00:00.000Z',
+      endDate: null,
+      signedDate: null,
+      fileReference: null,
+      description: null,
+      dateCreated: '2019-09-26T08:12:17.107Z',
+      dateUpdated: '2019-09-26T08:12:17.107Z',
+      agreementDescription: 'Section 127 (Two Part Tariff)'
+    }]
+  },
+  {
+    id: 'charge-element-3',
+    abstractionPeriod: {
+      startDay: 1,
+      startMonth: 11,
+      endDay: 31,
+      endMonth: 3
+    },
+    authorisedAnnualQuantity: '20',
+    season: 'winter',
+    source: 'unsupported',
+    loss: 'very low',
+    purposeUse: {
+      type: 'use',
+      code: '650',
+      name: 'Transfer Between Sources (Post Water Act 2003)'
+    },
+    factorsOverridden: false,
+    billableAnnualQuantity: null,
+    description: 'River Oak',
+    srocCategory: null,
+    dateCreated: '2019-09-26T08:12:08.346Z',
+    dateUpdated: '2019-09-26T08:12:08.346Z',
+    eiucSource: 'other',
+    chargeAgreements: []
+  },
+  {
+    id: 'charge-element-4',
+    abstractionPeriod: {
+      startDay: 1,
+      startMonth: 4,
+      endDay: 31,
+      endMonth: 10
+    },
+    authorisedAnnualQuantity: '22.7',
+    season: 'winter',
+    source: 'unsupported',
+    loss: 'high',
+    purposeUse: {
+      type: 'use',
+      code: '420',
+      name: 'Spray Irrigation - Direct'
+    },
+    factorsOverridden: true,
+    billableAnnualQuantity: '20',
+    description: 'Reservoir',
+    srocCategory: null,
+    dateCreated: '2019-09-26T08:12:08.346Z',
+    dateUpdated: '2019-09-26T08:12:08.346Z',
+    eiucSource: 'other',
+    chargeAgreements: [{
+      chargeAgreementId: 'charge-agreement-3',
+      id: 'charge-element-4',
+      agreementCode: 'S127',
+      startDate: '2017-10-11T23:00:00.000Z',
+      endDate: null,
+      signedDate: null,
+      fileReference: null,
+      description: null,
+      dateCreated: '2019-09-26T08:12:17.107Z',
+      dateUpdated: '2019-09-26T08:12:17.107Z',
+      agreementDescription: 'Section 127 (Two Part Tariff)'
+    }]
+  }];
 
 const returns = [{
   returnId: 'return-1',
@@ -676,6 +652,6 @@ const returns = [{
 }];
 
 module.exports = {
-  chargeVersion,
+  chargeElements,
   returns
 };
