+++ test/modules/billing/services/two-part-tariff-service/sample-data/sample-matching-data-6.js
@@ -1,125 +1,88 @@
-const chargeVersion1 = {
-  chargeVersionId: 'charge-version-1',
-  licenceRef: 'licence-ref',
-  scheme: 'alcs',
-  versionNumber: 108,
-  startDate: '2017-12-06T00:00:00.000Z',
-  status: 'current',
-  apportionment: false,
-  error: false,
-  endDate: null,
-  billedUptoDate: '2020-03-30T23:00:00.000Z',
-  regionCode: 1,
-  dateCreated: '2019-09-26T08:12:02.492Z',
-  dateUpdated: '2019-09-26T08:12:02.492Z',
-  source: 'nald',
-  chargeElements: [{
-    chargeElementId: 'charge-element-1',
-    chargeVersionId: 'charge-version-1',
-    abstractionPeriodStartDay: 1,
-    abstractionPeriodStartMonth: 4,
-    abstractionPeriodEndDay: 31,
-    abstractionPeriodEndMonth: 3,
-    startDate: '2017-12-06',
-    endDate: '2018-03-31',
-    totalDays: 365,
-    billableDays: 116,
-    authorisedAnnualQuantity: '112.2',
-    season: 'summer',
-    seasonDerived: 'summer',
-    source: 'unsupported',
-    loss: 'high',
-    purposePrimary: 'A',
-    purposeSecondary: 'AGR',
-    purposeTertiary: '420',
-    factorsOverridden: false,
-    billableAnnualQuantity: '56.1',
-    timeLimitedStartDate: null,
-    timeLimitedEndDate: null,
-    description: 'Big field',
-    srocCategory: null,
-    dateCreated: '2019-09-26T08:12:08.346Z',
-    dateUpdated: '2019-09-26T08:12:08.346Z',
-    purposePrimaryDescription: 'Agriculture',
-    purposeSecondaryDescription: 'General Agriculture',
-    purposeTertiaryDescription: 'Spray Irrigation - Storage',
-    eiucSource: 'other',
-    chargeAgreements: [{
-      chargeAgreementId: 'charge-agreement-1',
-      chargeElementId: 'charge-element-1',
-      agreementCode: 'S127',
-      startDate: '2017-12-06T00:00:00.000Z',
-      endDate: null,
-      signedDate: null,
-      fileReference: null,
-      description: null,
-      dateCreated: '2019-09-26T08:12:17.107Z',
-      dateUpdated: '2019-09-26T08:12:17.107Z',
-      agreementDescription: 'Section 127 (Two Part Tariff)'
-    }]
+const chargeElement1 = [{
+  id: 'charge-element-1',
+  abstractionPeriod: {
+    startDay: 1,
+    startMonth: 4,
+    endDay: 31,
+    endMonth: 3
+  },
+  startDate: '2017-12-06',
+  endDate: '2018-03-31',
+  totalDays: 365,
+  billableDays: 116,
+  authorisedAnnualQuantity: '112.2',
+  season: 'summer',
+  source: 'unsupported',
+  loss: 'high',
+  purposeUse: {
+    type: 'use',
+    code: '420',
+    name: 'Spray Irrigation - Storage'
+  },
+  factorsOverridden: false,
+  billableAnnualQuantity: '56.1',
+  description: 'Big field',
+  srocCategory: null,
+  dateCreated: '2019-09-26T08:12:08.346Z',
+  dateUpdated: '2019-09-26T08:12:08.346Z',
+  eiucSource: 'other',
+  chargeAgreements: [{
+    chargeAgreementId: 'charge-agreement-1',
+    id: 'charge-element-1',
+    agreementCode: 'S127',
+    startDate: '2017-12-06T00:00:00.000Z',
+    endDate: null,
+    signedDate: null,
+    fileReference: null,
+    description: null,
+    dateCreated: '2019-09-26T08:12:17.107Z',
+    dateUpdated: '2019-09-26T08:12:17.107Z',
+    agreementDescription: 'Section 127 (Two Part Tariff)'
   }]
-};
+}];
 
-const chargeVersion2 = {
-  chargeVersionId: 'charge-version-2',
-  licenceRef: 'licence-ref',
-  scheme: 'alcs',
-  versionNumber: 107,
-  startDate: '2014-04-13T23:00:00.000Z',
-  status: 'superseded',
-  apportionment: false,
-  error: false,
-  endDate: '2017-12-05T00:00:00.000Z',
-  billedUptoDate: '2017-12-05T00:00:00.000Z',
-  dateCreated: '2019-09-26T08:12:02.492Z',
-  dateUpdated: '2019-09-26T08:12:02.492Z',
-  source: 'nald',
-  chargeElements: [{
-    chargeElementId: 'charge-element-1',
-    chargeVersionId: 'charge-version-2',
-    abstractionPeriodStartDay: 1,
-    abstractionPeriodStartMonth: 1,
-    abstractionPeriodEndDay: 31,
-    abstractionPeriodEndMonth: 12,
-    startDate: '2017-04-01',
-    endDate: '2017-12-05',
-    totalDays: 365,
-    billableDays: 249,
-    authorisedAnnualQuantity: '112.2',
-    season: 'summer',
-    seasonDerived: 'summer',
-    source: 'unsupported',
-    loss: 'high',
-    purposePrimary: 'A',
-    purposeSecondary: 'AGR',
-    purposeTertiary: '400',
-    factorsOverridden: false,
-    billableAnnualQuantity: '112.2',
-    timeLimitedStartDate: null,
-    timeLimitedEndDate: null,
-    description: 'SEEPAGE RESV',
-    srocCategory: null,
-    dateCreated: '2019-09-26T08:12:08.346Z',
-    dateUpdated: '2019-09-26T08:12:08.346Z',
-    purposePrimaryDescription: 'Agriculture',
-    purposeSecondaryDescription: 'General Agriculture',
-    purposeTertiaryDescription: 'Spray Irrigation - Direct',
-    eiucSource: 'other',
-    chargeAgreements: [{
-      chargeAgreementId: 'charge-agreement-1',
-      chargeElementId: 'charge-element-1',
-      agreementCode: 'S127',
-      startDate: '2014-04-13T23:00:00.000Z',
-      endDate: null,
-      signedDate: null,
-      fileReference: null,
-      description: null,
-      dateCreated: '2019-09-26T08:12:17.107Z',
-      dateUpdated: '2019-09-26T08:12:17.107Z',
-      agreementDescription: 'Section 127 (Two Part Tariff)'
-    }]
+const chargeElement2 = [{
+  id: 'charge-element-1',
+  abstractionPeriod: {
+    startDay: 1,
+    startMonth: 1,
+    endDay: 31,
+    endMonth: 12
+  },
+  startDate: '2017-04-01',
+  endDate: '2017-12-05',
+  totalDays: 365,
+  billableDays: 249,
+  authorisedAnnualQuantity: '112.2',
+  season: 'summer',
+  source: 'unsupported',
+  loss: 'high',
+  purposeUse: {
+    type: 'use',
+    code: '400',
+    name: 'Spray Irrigation - Direct'
+  },
+  factorsOverridden: false,
+  billableAnnualQuantity: '112.2',
+  description: 'SEEPAGE RESV',
+  srocCategory: null,
+  dateCreated: '2019-09-26T08:12:08.346Z',
+  dateUpdated: '2019-09-26T08:12:08.346Z',
+  eiucSource: 'other',
+  chargeAgreements: [{
+    chargeAgreementId: 'charge-agreement-1',
+    id: 'charge-element-1',
+    agreementCode: 'S127',
+    startDate: '2014-04-13T23:00:00.000Z',
+    endDate: null,
+    signedDate: null,
+    fileReference: null,
+    description: null,
+    dateCreated: '2019-09-26T08:12:17.107Z',
+    dateUpdated: '2019-09-26T08:12:17.107Z',
+    agreementDescription: 'Section 127 (Two Part Tariff)'
   }]
-};
+}];
 
 const returns = [{
   returnId: 'return-1',
@@ -371,7 +334,7 @@ const returns = [{
 }];
 
 module.exports = {
-  chargeVersion1,
-  chargeVersion2,
+  chargeElement1,
+  chargeElement2,
   returns
 };
