+++ test/modules/billing/services/two-part-tariff-service/test-charge-data.js
@@ -7,35 +7,26 @@ const defaults = {
 
 const createChargeElement = (opts = {}) => {
   const options = Object.assign({}, defaults, opts);
-  return {
-    chargeElementId: options.chargeElementId || 'charge-element-id',
-    chargeVersionId: 'charge-version-id',
-    abstractionPeriodStartDay: options.abstractionPeriodStartDay,
-    abstractionPeriodStartMonth: options.abstractionPeriodStartMonth,
-    abstractionPeriodEndDay: options.abstractionPeriodEndDay,
-    abstractionPeriodEndMonth: options.abstractionPeriodEndMonth,
+  const chargeElement = {
+    id: options.chargeElementId || 'charge-element-id',
+    abstractionPeriod: {
+      startDay: options.abstractionPeriodStartDay,
+      startMonth: options.abstractionPeriodStartMonth,
+      endDay: options.abstractionPeriodEndDay,
+      endMonth: options.abstractionPeriodEndMonth
+    },
     actualReturnQuantity: options.actualReturnQuantity,
     authorisedAnnualQuantity: options.authorisedAnnualQuantity,
     season: options.season,
-    seasonDerived: options.seasonDerived,
     source: options.source,
     loss: options.loss,
-    purposePrimary: 'A',
-    purposeSecondary: 'AGR',
-    purposeTertiary: options.purposeTertiary,
-    factorsOverridden: false,
     billableAnnualQuantity: options.billableAnnualQuantity,
-    timeLimitedStartDate: options.timeLimitedStartDate,
-    timeLimitedEndDate: options.timeLimitedEndDate,
     description: 'Borehole-NGR XX 123 456. Summer Spray Irrigation.',
     srocCategory: null,
     dateCreated: '2019-08-15T04:43:45.234Z',
     dateUpdated: '2019-09-27T10:23:07.981Z',
     proRataAuthorisedQuantity: options.proRataAuthorisedQuantity,
     maxPossibleReturnQuantity: options.maxPossibleReturnQuantity,
-    purposePrimaryDescription: 'Agriculture',
-    purposeSecondaryDescription: 'General Agriculture',
-    purposeTertiaryDescription: 'Spray Irrigation - Direct',
     startDate: options.startDate,
     endDate: options.endDate,
     // Number of days this charge element would have been billable for
@@ -46,22 +37,22 @@ const createChargeElement = (opts = {}) => {
     // account the abstraction period and the date range of the element
     billableDays: options.billableDays
   };
-};
 
-const wrapElementsInVersion = (chargeElements, startDate, endDate) => {
-  return {
-    chargeVersion: {
-      licenceRef: '6/33/53/*S/0153'
-    },
-    financialYear: {
-      startDate: '2016-04-01',
-      endDate: '2017-03-31'
-    },
-    startDate: startDate,
-    endDate: endDate,
-    chargeElements: chargeElements
-  };
+  if (options.purposeTertiary) {
+    chargeElement.purposeUse = {
+      type: 'use',
+      code: options.purposeTertiary,
+      name: 'Spray Irrigation - Direct'
+    };
+  }
+  if (options.timeLimitedStartDate || options.timeLimitedEndDate) {
+    chargeElement.timeLimitedPeriod = {
+      startDate: options.timeLimitedStartDate,
+      endDate: options.timeLimitedEndDate
+    };
+  }
+
+  return chargeElement;
 };
 
 exports.createChargeElement = createChargeElement;
-exports.wrapElementsInVersion = wrapElementsInVersion;
